Fixed short conversions
git-svn-id: http://svn.littleshoot.org/svn/littleshoot/trunk/common/stun/stack@430 8d7aff80-242d-0410-bcee-ed6f15385b3d

+++ src/main/java/org/lastbamboo/common/stun/stack/message/attributes/StunAttributeWriter.java
@@ -29,8 +29,8 @@ public class StunAttributeWriter implements StunAttributeVisitor
 
     public void visitMappedAddress(final MappedAddress address)
         {
-        final short type = 
-            StunAttributeType.MAPPED_ADDRESS.convert().shortValue();
+        final int type = 
+            StunAttributeType.MAPPED_ADDRESS.convert().intValue();
         final int length = address.getBodyLength();
         m_buf.putShort((short) (type & 0xffff));
         m_buf.putShort((short) (length & 0xffff));
